fix: make corrections as per the criteria
