+++ src/server/points-api-service.js
@@ -1,4 +1,4 @@
-import {Method} from '../utils/const.js';
+import {Method} from '../utils/enum.js';
 
 import ApiService from '../framework/api-service.js';
