+++ src/utils/utils.js
@@ -1,9 +1,7 @@
-const getRandomInteger = (min = 1, max = 100) => Math.floor(Math.random() * (max - min + 1)) + min;
-
-const upFirstLetter = (word) => `${word[0].toUpperCase()}${word.slice(1)}`;
+const capitalizeFirstLetter = (word) => `${word[0].toUpperCase()}${word.slice(1)}`;
 
 const sortByValue = (value) => (first, second) => second[value] - first[value];
 
 const updateItem = (items, update) => (items.map((item) => item.id === update.id ? update : item));
 
-export {getRandomInteger, upFirstLetter, sortByValue, updateItem};
+export {capitalizeFirstLetter, sortByValue, updateItem};
