+++ src/view/toolbar/cta-button-view.js
@@ -3,20 +3,20 @@ import AbstractView from '../../framework/view/abstract-view.js';
 const createCtaButtonTemplate = () => '<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>';
 
 export default class CtaButtonView extends AbstractView {
-  #handleClick = null;
+  #handleCtaButtonClick = null;
 
-  constructor({onClick}) {
+  constructor({onCtaButtonClick}) {
     super();
-    this.#handleClick = onClick;
-    this.element.addEventListener('click', this.#clickHandler);
+    this.#handleCtaButtonClick = onCtaButtonClick;
+    this.element.addEventListener('click', this.#ctaButtonClickHandler);
   }
 
   get template() {
     return createCtaButtonTemplate();
   }
 
-  #clickHandler = (evt) => {
+  #ctaButtonClickHandler = (evt) => {
     evt.preventDefault();
-    this.#handleClick();
+    this.#handleCtaButtonClick();
   };
 }
